Evaluation

Simulated data

LatentDynamics.eval_prediction — Function

eval_prediction(m::odevae, data::simdata, n_future_tps::Int=1)

Evaluate the prediction performance of the model on the latent representations.

Arguments

m::odevae: trained odevae model

testdata: simdata object containing the simulated data

n_future_tps::Int=1: number of future time points to predict

Returns

Vectors of prediction errors of different models for each patient (summed over all time points):

ODEprederrs: vector of prediction errors for the ODE model

OLSprederrs: vector of prediction errors for an OLS model, also fitted in a piecewise manner similar to the ODE model

interceptprederrs: vector of prediction errors of an intercept model

locfprederrs: vector of prediction errors of a LOCF model

LatentDynamics.eval_reconstructed_prediction — Function

eval_reconstructed_prediction(m::odevae, testdata::SMATestData, n_future_tps::Int=1)

Evaluate the prediction performance of the model using the reconstructed data obtained from the decoder, based on simulated data.

Arguments

m::odevae: trained odevae model

testdata: simdata object containing the simulated data

n_future_tps::Int=1: number of future time points to predict

Returns

Vectors of prediction errors of different models for each patient (summed over all time points):

ODEprederrs: vector of prediction errors for the ODE model

OLSprederrs: vector of prediction errors for an OLS model, also fitted in a piecewise manner similar to the ODE model

interceptprederrs: vector of prediction errors of an intercept model

locfprederrs: vector of prediction errors of a LOCF model

SMArtCARE data

LatentDynamics.eval_prediction — Function

eval_prediction(m::odevae, testdata::SMATestData, n_future_tps::Int=1)

Evaluate the prediction performance of the model on the latent representations.

Arguments

m::odevae: trained odevae model

testdata: SMATestData object containing the SMArtCARE data

n_future_tps::Int=1: number of future time points to predict

Returns

Vectors of prediction errors of different models for each patient (summed over all time points):

ODEprederrs: vector of prediction errors for the ODE model

OLSprederrs: vector of prediction errors for an OLS model, also fitted in a piecewise manner similar to the ODE model

interceptprederrs: vector of prediction errors of an intercept model

locfprederrs: vector of prediction errors of a LOCF model

LatentDynamics.eval_reconstructed_prediction — Function

eval_reconstructed_prediction(m::odevae, testdata::SMATestData, n_future_tps::Int=1)

Evaluate the prediction performance of the model using the reconstructed data obtained from the decoder.

Arguments

m::odevae: trained odevae model

testdata: SMATestData object containing the SMArtCARE data

n_future_tps::Int=1: number of future time points to predict

Returns

Vectors of prediction errors of different models for each patient (summed over all time points):

ODEprederrs: vector of prediction errors for the ODE model

OLSprederrs: vector of prediction errors for an OLS model, also fitted in a piecewise manner similar to the ODE model

interceptprederrs: vector of prediction errors of an intercept model

locfprederrs: vector of prediction errors of a LOCF model