Plotting

We provide various functions to visualize the learnt latent trajectories, both for individual patients and panels of multiple patients, for both simulated data and SMArtCARE data. We provide functions to plot the trajectories of the overall estimator of the solution and functions to plot the prediction from the model starting at the current time point until the next time point (the _piecewise functions).

Simulated data

LatentDynamics.plot_truesolution — Function

plot_truesolution(group, 
    data::simdata, 
    t_range, 
    sol_group1, 
    sol_group2; 
    showdata=true
)

Plot the true solution of the ODE system for one simulated groups of patients.

Arguments

group: number of the group of patients whose true underlying trajectory to plot (1 or 2)

data: simdata object containing the simulated data

t_range: range of time points to plot

sol_group1: true solution of the ODE system for group 1

sol_group2: true solution of the ODE system for group 2

Optional keyword arguments

showdata: whether to show the data points (default: true)

Individual trajectories

LatentDynamics.createindividualplot — Method

createindividualplot(m::odevae, 
    data::simdata, 
    idx::Int, 
    sol::Matrix, 
    trange, 
    args::LossArgs;
    title::String="", 
    showtruesol::Bool=true,
    axislabs::Bool=false, 
    showOLS::Bool=true, 
    colors_truesol::Array{String} = ["#ff7f0e" "#1f77b4"]
)

Create a plot of the latent representation of a single patient based on simulated data, using the overall ODE solution obtained as a (weighted) average of the individual ODE solutions (based on using each observed time point as initial condition).

Arguments

m: trained odevae model

data: simdata object containing the simulated data

idx: index of the patient whose latent representation to plot

sol: true solution of the ODE system

trange: range of time points to plot

args: LossArgs object containing the arguments controlling the loss function behaviour (and hence the details of how the latent trajectories are computed)

Optional keyword arguments

title: title of the plot (default: "")

showtruesol: whether to show the true solution of the ODE system (default: true)

axislabs: whether to show the axis labels (default: false)

showOLS: whether to show the OLS fit to the latent representation (default: true)

colors_truesol: colors to use for the true solution of the ODE system (default: ["#ff7f0e" "#1f77b4"])

LatentDynamics.createindividualplot_piecewise — Method

createindividualplot_piecewise(m::odevae, 
    data::simdata, 
    ind::Int, 
    sol::Matrix, 
    t_range; 
    title::String="", 
    axislabs::Bool=false, 
    showtruesol::Bool=true,
    showOLS::Bool=true, 
    showglobalOLS::Bool=false,
    colors_truesol::Array{String} = ["#ff7f0e" "#1f77b4"]
)

Create a plots of the latent representation of a individual simulated patient, using the individual ODE solutions from each timepoint as the initial condition to the next, hence creating a 'piecewise' trajectory.

Arguments

m: odevae model

data: simdata object containing the simulated data

ind: index of the patient whose latent representation to plot

sol: true solution of the ODE system for the patient

t_range: range of time points to plot

Optional keyword arguments

title: title of the plot (default: "")

axislabs: whether to show the axis labels (default: false)

showtruesol: whether to show the true solution of the ODE system (default: true)

showOLS: whether to show the piecewise OLS fit to the latent representation (default: true)

showglobalOLS: whether to show a global OLS fit to the latent representation using all timepoints (default: false)

colors_truesol: colors to use for the true solution and the ODE solution (default: ["#ff7f0e" "#1f77b4"])

LatentDynamics.eval_z_trajectories — Function

eval_z_trajectories(m::odevae, 
    data::simdata, 
    inds::Array{Int}, 
    sol_group1::Matrix, 
    sol_group2::Matrix, 
    t_range, 
    args::LossArgs; 
    title::String="", 
    showtruesol::Bool=true,
    axislabs::Bool=false, 
    showOLS::Bool=true, 
    swapcolorcoding::Bool=false
)

Create a plot of the latent representation of a set of patients based on simulated data, using the overall ODE solution obtained as a (weighted) average of the individual ODE solutions (based on using each observed time point as initial condition). Used as a callback during training of the odevae model.

Arguments

m: odevae model

data: simdata object containing the simulated data

inds: array of indices of the patients whose latent representations to plot

sol_group1: true solutions of the ODE system for patients in group 1

sol_group2: true solutions of the ODE system for patients in group 2

t_range: range of time points to plot

args: LossArgs object containing the arguments controlling the loss function behaviour (and hence the details of how the latent trajectories are computed)

Optional keyword arguments

title: title of the plot (default: "")

showtruesol: whether to show the true solution of the ODE system (default: true)

axislabs: whether to show the axis labels (default: false)

showOLS: whether to show the OLS fit to the latent representation (default: true)

swapcolorcoding: whether to swap the color coding of the true solutions in the two groups (default: false)

Panels of trajectories

Missing docstring.

Missing docstring for plot_seleced_ids_piecewise(m::odevae, data::simdata, inds::Array{Int}, sol_group1::Matrix, sol_group2::Matrix, t_range; axislabs::Bool=false, showtruesol::Bool=true, showtitle::Bool=true, showOLS::Bool=true, showglobalOLS::Bool=false, swapcolorcoding::Bool=false, sort_inds::Bool=true ). Check Documenter's build log for details.

SMArtCARE data

Individual trajectories

LatentDynamics.createindividualplot — Method

createindividualplot(m::odevae, 
    testdata::SMATestData, 
    args::LossArgs, 
    patient_id; 
    axislabs::Bool=false, 
    title::String="", 
    showOLS::Bool=true
)

Create a plot of the latent representation of a single patient from a trained odevae model, using the overall ODE solution obtained as a (weighted) average of the individual ODE solutions (based on using each observed time point as initial condition).

Arguments

m: trained odevae model

testdata: SMATestData object containing the data

args: LossArgs object containing the arguments controlling the loss function behaviour (and hence the details of how the latent trajectories are computed)

patient_id: ID of the patient whose latent representation to plot

Optional keyword arguments

axislabs: whether to add axis labels to the plot (default: false)

title: title of the plot (default: "")

showOLS: whether to show the OLS fit to the latent representation (default: true)

LatentDynamics.createindividualplot_piecewise — Method

createindividualplot_piecewise(m::odevae, 
    testdata::SMATestData, 
    patient_id; 
    title::String="", 
    showOLS::Bool=true, 
    axislabs::Bool=false
)

Create a plot of the latent representation of a single patient from a trained odevae model, using the individual ODE solutions from each timepoint as the initial condition to the next, hence creating a 'piecewise' trajectory.

Arguments

m: trained odevae model

testdata: SMATestData object containing the data

patient_id: ID of the patient whose latent representation to plot

Optional keyword arguments

axislabs: whether to add axis labels to the plot (default: false)

title: title of the plot (default: "")

showOLS: whether to show the OLS fit to the latent representation (default: true)

Panels of trajectories

LatentDynamics.plot_selected_ids — Function

plot_selected_ids(m::odevae, 
    testdata::SMATestData, 
    args::LossArgs, 
    selected_ids::Array; 
    showOLS::Bool=true
)

Create a panel of plots of the latent representations of a list of patients, using the overall ODE solution obtained as a (weighted) average of the individual ODE solutions (based on using each observed time point as initial condition).

Arguments

m: trained odevae model

testdata: SMATestData object containing the data

args: LossArgs object containing the arguments controlling the loss function behaviour (and hence the details of how the latent trajectories are computed)

selected_ids: array of the IDs of the patients whose latent representations to plot

Optional keyword arguments

showOLS: whether to show the OLS fit to the latent representation (default: true)

LatentDynamics.plot_selected_ids_piecewise — Method

plot_selected_ids_piecewise(m::odevae, 
    testdata::SMATestData,
    selected_ids::Array; 
    showOLS::Bool=true
)

Create a panel of plots of the latent representations of a list of patients, using the individual ODE solutions from each timepoint as the initial condition to the next, hence creating a 'piecewise' trajectory.

Arguments

m: trained odevae model

testdata: SMATestData object containing the data

selected_ids: array of the IDs of the patients whose latent representations to plot

Optional keyword arguments

showOLS: whether to show the OLS fit to the latent representation (default: true)